readme

+++ homework/hw2/code/constants.h
@@ -1,6 +1,6 @@
 #ifndef CSA_HWS_CONSTANTS_H
 #define CSA_HWS_CONSTANTS_H
 
-const size_t MAX_NAME_LEN = 40;
+const int MAX_NAME_LEN = 40;
 
 #endif
